Refactor: Move Google client info to config
This commit moves the Google client information from the code to a configuration file.

The `GoogleClientInfo` class is now loaded from the configuration file, making it easier to manage and update the client information without recompiling the code.

The `clientInfo` field has been removed from the `CreateGmailAccountRequest` class, as it is no longer needed to be passed in the request.

The `AuthConfig` class has been simplified, as it no longer needs to hold the client information.

The `GmailSupport` class has been updated to use the `GoogleClientInfo` from the configuration file.

The `GmailPool` class has also been updated to use the `GoogleClientInfo` from the configuration file when loading the account state.

The tests have been updated to reflect these changes.